Refactor: 모듈화 완료
