+++ client/front/lib/status/status_slide/status_runner.dart
@@ -6,14 +6,14 @@ import 'package:http/http.dart' as http;
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'custom_dialog.dart';
-import 'main/errand_list/errand_list.dart';
+import '../../widgets/dialog/custom_dialog.dart';
+import '../../main/errand_list/errand_list.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:stomp_dart_client/stomp_dart_client.dart';
 import 'package:dropdown_button2/dropdown_button2.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'map.dart';
-import 're-show_errand.dart';
+import '../status_icons/map.dart';
+import '../status_icons/re-show_errand.dart';
 final storage = FlutterSecureStorage();
 class StatusContent{//진행중인 심부름이 간략하게 담고 있는 정보들
   String contents; //메시지
