+++ client/front/lib/widgets/dialog/custom_dialog.dart
